Add files via upload
Matrix Multiplication Problem

Given a sequence of matrices, find the most efficient way to multiply these matrices together.
The problem is not actually to perform the multiplications, but merely to decide in which order 
to perform the multiplications. We have many options to multiply a chain of matrices because 
matrix multiplication is associative. In other words, no matter how we parenthesize the product, 
the result will be the same

Input1:
4
10 20 30 40 30
Output1:
30000

Input2:
2
10 20 30
Output2:
6000